replace march flag with adx for arch agnostic builds
in the go binding, the march=native flag was causing panics when built on one architecture
and ran on another.
for example compiling blst on a system with AMD architecture and running
binary on an Intel system would result in a panic.
blst would be built with the sha-ni extensions on that Ryzen platform
that then wouldn't have been available on that Intel processor.
using the ADX flag will work on Intel/AMD processor from the last ~ 5 years